requirements.txt
mlflow
dvc
torch
openai-whisper
TTS
pyyaml
